requirements.txt
arrow==0.12.1
click==6.7
Flask==1.0.2
gunicorn==19.9.0
itsdangerous==0.24
Jinja2==2.10
MarkupSafe==1.0
python-dateutil==2.7.3
python-decouple==3.1
redis==2.10.6
rq==0.12.0
rq-dashboard==0.3.12
six==1.11.0
Werkzeug==0.14.1
